+++ demo-app/package.json
@@ -2,7 +2,7 @@
   "name": "demo-app",
   "version": "0.1.0",
   "devDependencies": {
-    "app-framework": "^0.30.3"
+    "app-framework": "^0.31.0"
   },
   "scripts": {
     "dev": "cd node_modules/app-framework && node build/dev-server.js",
@@ -13,6 +13,7 @@
   "title": "Demo App",
   "theme": "ios",
   "defaultLanguage": "en",
+  "showFrameOnDesktop": true,
   "routes": {
     "home": "home",
     "panelLeft": "panelLeft",
