+++ demo-app/pages/popup.vue
@@ -12,6 +12,9 @@
         <f7-input v-model="text" placeholder="Some text ..." type="text"></f7-input>
       </f7-list-item>
     </f7-list>
+    <f7-block style="text-align: center">
+      <f7-link close-popup>Close popup</f7-link>
+    </f7-block>
   </f7-page>
 </template>
