+++ index.ejs
@@ -8,6 +8,8 @@
     <title><%= htmlWebpackPlugin.options.title %></title>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="frame">
+      <div id="app"></div>     
+    </div>
   </body>
 </html>
