+++ main.css
@@ -6,3 +6,58 @@
 .framework7-root {
   visibility: hidden;
 }
+body {
+  background: #F2F2F2;
+}
+#frame {
+  width: 100%;
+  height: 100%;
+}
+.phone {
+  position: absolute;
+  top: calc((100% - 568px - 210px) / 2);
+  left: calc((100% - 320px - 50px) / 2);
+  height: 568px !important;
+  width: 320px !important;
+  background: #111;
+  border-radius: 55px;
+  box-shadow: 0px 0px 0px 2px #aaa;
+  padding: 105px 25px;
+  transition: 0.4s;
+}
+.phone:after {
+  content: '';
+  position: absolute;
+  width: 60px;
+  height: 60px;
+  left: 50%;
+  margin-left: -30px;
+  bottom: 20px;
+  border-radius: 100%;
+  box-sizing: border-box;
+  border: 5px solid #333;
+}
+.phone:before {
+  position: absolute;
+  left: 50%;
+  margin-left: -30px;
+  top: 50px;
+  content: '';
+  background: #333;
+  width: 60px;
+  height: 10px;
+  border-radius: 10px;
+  transition: 0.4s;
+}
+.limitWidth {
+  position: absolute;
+  left: calc((100% - 320px) / 2) !important;
+  width: 320px !important;
+  transition: 0.4s;
+}
+.limitHeight {
+  position: absolute;
+  top: calc((100% - 568px) / 2) !important;
+  height: 568px !important;
+  transition: 0.4s;
+}
